test accurary before recalibration - step 1, 2

+++ app/src/main/java/com/calorieko/app/ble/BleScaleManager.kt
@@ -57,8 +57,8 @@ class BleScaleManager(private val context: Context) {
     private val _calibrationEvent = MutableSharedFlow<String>(replay = 1, extraBufferCapacity = 1)
     val calibrationEvent: SharedFlow<String> = _calibrationEvent
 
-    private val _liveWeight = MutableStateFlow(0)
-    val liveWeight: StateFlow<Int> = _liveWeight.asStateFlow()
+    private val _liveWeight = MutableStateFlow(0f)
+    val liveWeight: StateFlow<Float> = _liveWeight.asStateFlow()
 
     private val bluetoothAdapter: BluetoothAdapter? =
         (context.getSystemService(Context.BLUETOOTH_SERVICE) as? BluetoothManager)?.adapter
@@ -284,7 +284,7 @@ class BleScaleManager(private val context: Context) {
                 _calibrationEvent.tryEmit(strValue.trim())
             } else if (characteristic.uuid == WEIGHT_CHAR_UUID) {
                 val weightFloat = strValue.trim().toFloatOrNull() ?: 0f
-                _liveWeight.value = weightFloat.toInt()
+                _liveWeight.value = weightFloat
                 Log.d(TAG, "Weight received: $strValue (parsed: ${_liveWeight.value}g)")
             } else if (characteristic.uuid == COMMAND_CHAR_UUID) {
                 Log.d(TAG, "Command response: $strValue")
